# backend/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "NYDAX API",
    "description": "NYDAX digital asset exchange endpoint."
  },
  "host": "ixp.nydax.com:2083",
  "basePath": "/v1",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/transaction": {
      "get": {
        "tags": ["Transaction"],
        "summary": "Get Transaction history of the user.",
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "object",
              "properties": {
                "Authorization": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "500": {
            "description": "ERROR",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          }
        }
      },
      "post": {
        "tags": ["Transaction"],
        "summary": "Get Transaction history of the user.",
        "parameters": [
          {
            "name": "Authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "object",
              "properties": {
                "Authorization": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "Transaction",
            "required": true,
            "in": "body",
            "type": "object"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessMessage"
            }
          },
          "500": {
            "description": "ERROR",
            "schema": {
              "$ref": "#/definitions/ErrorMessage"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorMessage": {
      "type": "object",
      "properties": {
        "error": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            }
          }
        }
      }
    },
    "SuccessMessage": {
      "type": "object",
      "properties": {
        "message": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "description": {
              "type": "string"
            }
          }
        }
      }
    },
...
  }
}